Move Layout and navbar component to web workspace
This components are specific to web and should therefore be in this workspace
