+++ apps/web/components/__tests__/Navbar.test.js
