+++ apps/web/pages/place-picker.js
@@ -3,7 +3,8 @@
 import * as React from 'react';
 import { View } from 'react-native';
 import { PlacePicker } from '@kiwicom/margarita-core';
-import { Layout } from '@kiwicom/margarita-components';
+
+import Layout from '../components/Layout';
 
 export default () => (
   <Layout>
