+++ packages/components/index.js
@@ -1,6 +1,5 @@
 // @flow
 
-export { default as Layout } from './src/Layout';
 export {
   default as TouchableWithoutFeedback,
 } from './src/TouchableWithoutFeedback';
